Fix handling of SSL errors
Most SSL errors represent failures in establising a secure connection.
These cases should only be reported instead of entering the debugger.

In ACL the error occurs much later in the handling of a request (due
to a lazy approach in establishing a connection with the client).
In SBCL the error occurs when creating a SSL socket.

Tests performed:
 - Build web-server, started the web-server with client authorization
 via certificates and check for errors when connecting from a client
 with no certificates. As expected, the connection is refused however,
 no error occurs.
 - If a client provides a valid certificate the connection is
 established (although an error is raised in the application level,
 that will be handled in the appropriate place: SCS).

Issue: PSCS-360
Reference: POA.37293.0
Change-Id: Id79e00dae5d14bf380399796ae9f730bec6f5350
Reviewed-on: https://gerrit.siscog/c/open-source/zaserve/+/6592
Reviewed-by: Rui Aurélio <rui.aurelio@siscog.pt>