Fix definition of BASE64-ENCODE
ZASERVE is loaded using a special reader which contains features to
emulate the use of Allegro Common Lisp.

While compiling this file in SBCL, the feature :ALLEGRO exists,
causing the wrong version of this function to compile.
Additionaly, as ZASERVE loads the original ASERVE library when in
ACL, we can safely discard the ACL specific version of this code
because it will never be loaded in ACL.

Issue: PSCS-479
Reference: POA.37293.0
Change-Id: Ie6e5652159d3d43433d1cc41234eb75d11e179e9
Reviewed-on: https://gerrit.siscog/c/open-source/zaserve/+/6749
Reviewed-by: Rui Aurélio <rui.aurelio@siscog.pt>